feat: conditionally nuke the current build

+++ run.bash
@@ -1,8 +1,12 @@
 #!/bin/bash
-rm -rf build
-mkdir build && cd build
+if [[ "$1" == "root" ]]; then
+  rm -rf build
+  mkdir build && cd build
 
-cmake ..
-cmake --build .
+  cmake ..
+  cmake --build .
+else
+  cd build
+fi
 
 ./unit-conversion
